autoposter-backend/requirements.txt
watchdog==6.0.0
python-dotenv==1.1.1
instagrapi==2.2.1
Pillow==11.3.0
tweepy==4.16.0
schedule==1.2.2
moviepy==1.0.3
cryptography==41.0.7
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# Multi-platform social media APIs
linkedin-api==2.2.0
google-api-python-client==2.110.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
facebook-sdk==3.1.0
TikTokApi==6.1.0

# Additional utilities
aiofiles==23.2.1
asyncio==3.4.3
psycopg2-binary==2.9.9
redis==5.0.1
